Allow overriding Authorization header prefix
Seeing this with the SAP Oauth 2.0 service. It returns a token with type "BearerToken".

This token type is then re-used by golang.org/x/oauth2 as prefix when sending the authenticated requests.

Unfortunately, the service in question does only accept Authorization header prefix "Bearer" (as we're used to).

So allow to configure the http client to override this.

Intended example usage:

```golang
    opts := []httputils.HTTPClientConfigOpt{httputils.LogCalls(logging.Call)}

	opts = append(opts, httputils.UseOAuth2(setTokenSource(cfg.Username, cfg.Password, cfg.OAuthTokenURL)))
	opts = append(opts, httputils.OverrideAuthHeaderPrefix("Bearer"))

	return &restServiceSink{
		endpoint: cfg.Endpoint,
		client:   httputils.NewHTTPClient(opts...),
	}
```